fix(db): keep the aim fingerprint integral end to end
round_id_sum stored SUM(round_id) — a bigint in Postgres, since round_id is an
integer — in a NUMERIC column, which the code then converted to float. Nothing is
wrong with today's magnitudes (~1.1e9 against a 2^53 mantissa), but tying a
cache's validity to mantissa width for no gain is a latent hazard, and the typing
misrepresented the data. Column is now BIGINT and the value stays int in Python,
so the comparison is exact integer equality.

Verified on dev after a clean re-apply: column reports bigint, the fingerprint
comes back as (int, int, int), a read after a write hits, and bumping
round_id_sum by one misses. Migration ledger validation CLEAN.

Both from copilot review on #746. The bootstrap dump carries the same change so
dump ≡ migrations still holds.

+++ migrations/077_player_aim_summary.sql
@@ -20,6 +20,9 @@
 --   last_event_time new shots, even in the pathological equal-count case
 --   round_id_sum    re-linking, which changes the flick index because that
 --                   window is partitioned by round even when no shot moved
+-- SUM(round_id) over an integer column is a bigint, so the column is BIGINT and
+-- the comparison is exact integer equality — a float round-trip would put the
+-- freshness of the cache at the mercy of mantissa width.
 -- A read compares the stored fingerprint with a fresh one; anything different
 -- means recompute. There is no TTL to tune and no import hook to forget.
 --
@@ -35,7 +38,7 @@ CREATE TABLE IF NOT EXISTS player_aim_summary (
     formula_version  INTEGER     NOT NULL,
     shot_count       BIGINT      NOT NULL,
     last_event_time  BIGINT      NULL,
-    round_id_sum     NUMERIC     NULL,
+    round_id_sum     BIGINT      NULL,
     payload          JSONB       NOT NULL,
     computed_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
